+++ js/modules/image-loader.js
@@ -1,0 +1,7 @@
+import "./image-loader/comments.js";
+import "./image-loader/form-clear.js";
+import "./image-loader/form-validator.js";
+import "./image-loader/hash-tags.js";
+import "./image-loader/image-effects.js";
+import "./image-loader/image-scale.js";
+import "./image-loader/post-image.js";
