+++ js/modules/random-content.js
@@ -1,0 +1,80 @@
+import {getRandomNumber, getRandomArrayElement, getRandomUniqueId} from "./utils.js";
+
+
+const USERS_NAMES = [
+  'Слава',
+  'Коля',
+  'Петя',
+  'Даша',
+  'Глаша',
+  'Та самая',
+  'BMW_1988',
+  'Artem_superman_92',
+  'Д.Жуан',
+  'Жора',
+  'ТВОЯ',
+  'Не твоя',
+  'БМВова',
+  'Боря',
+];
+
+const PHOTO__DESCRIPTION = [
+  'Мечта интроверта - пустой пляж. А кто спасать будет, если начнешь тонуть',
+  'Go to thr bitch',
+  'Я тут! Завидуете?',
+  'Дай подержать камеру и сфоткай меня на телефон',
+  'Суп с котом',
+  'Мои понты!',
+  'Мои завтраки в предверии пляжного сезона',
+  'Жду когда перебродит',
+  'Эй,забери меня домой, я уже обгорела',
+  'Как научить домочадцев ставить обувь на место? Подробности в по ссылке в профиле',
+  'Услуги по монтажу заборов - +7 495 45 45 45',
+  'Фото до прозда лежачего полицейского',
+  'Угадайте, это рыба или арбуз?',
+  'домкдлоикмдшг - по японски суши с котом (значит то же самое что у нас суп с котом',
+  'Если отключили отопление и выбило окно - тапочки для дома с подогревом - самое то',
+  'лети лети лепесток....',
+  'ОЙ, как я тут оказалась?!',
+  'ВСе потратил на машину, на ремонт денег не осталось',
+  'для ночных бдений',
+  'Лысые пальмы - свидетельствуют о том, что район ветренный',
+  'Бурда с лемоном',
+  'Хочу сюда!',
+  'Сегодня видела такого - ненавижу пауков',
+  'Я наконец попала на концерт мистера Кастрюля! Завидуйте!',
+  'Мой бегемотик и не мой бегемотик',
+];
+
+const MESSAGE_CONTENT = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+
+
+function getRandomCommentsArray() {
+return  Array(getRandomNumber(1, 9)).fill().map (() => ({
+  id: getRandomUniqueId(),
+  avatar: `img/avatar-${  getRandomNumber(0, 5)  }.svg`,
+  message: getRandomArrayElement(MESSAGE_CONTENT),
+  name: getRandomArrayElement(USERS_NAMES),
+}))
+}
+
+
+function getRandomComtent() {
+return Array(25).fill().map ((value,index) => ({
+  id: index+1,
+  url: `photos/${  index+1  }.jpg`,
+  description: PHOTO__DESCRIPTION[index],
+  likes: getRandomNumber(0, 1000),
+  comments: getRandomCommentsArray(),
+}))
+}
+
+export  {getRandomComtent};
